Emit a \ line continuation when one is used between string literals in original source
Ruby allows string literals to be continued on another line like this:

    "a string" \
    "another"

When parser sees this, it generates 2 (str "...") nodes, one after the other.
Therefore, make Unparser::Emitter::Literal::DynamicBody generate code which uses
a line continuation when it sees 2 adjacent str nodes under a dstr node, which
were separated by a line break in the original source.

This makes the 'corpus' integration test pass on Mutant. It didn't before.