Don't use parentheses around operator expressions when it is obvious they aren't needed
To do this, we have to take operator precedence and associativity into account,
to make sure we don't change the meaning of an expression by omitting the parens.
More then that, we need to consider what precedence rules are "obvious" to an
average Ruby programmer. In many cases, parens are not strictly necessary,
but may make the meaning clearer.

On the other hand, if we *always* emit parens, the code becomes littered with
parens which are obviously redundant, and that sucks.